fzr: remove header-lines
If there's no output left after the header lines, fzf
thinks there was no match, so it returns an error code.

Another option if I really wanted to make the input display as
a header line would be to catch the "no match" error code (1)
and treat that as success. I think it's fine the way it is.