wrappers: absolutely overengineer the crap out of meson
Now the meson wrapper understands `-Dcache=(none|ccache|sccache)` (plus
any commonly used variations of `false`), plus `-Dccache` and `-Dsccache`
(with an optional assignment of any commonly used variation of `true`),
in which case the specified compiler wrapper will be used.

Otherwise, the wrapper that's in $PATH will be used (if any).

In either case, meson's built-in autodetection will be disabled.