easyrsa: init-pki: actually find vars.example outside of $EASYRSA_PKI
As implemented, install_data_to_pki() will never install a vars file
if it does not already exist in $EASYRSA_PKI (and the entire premise of
this function is to allow creating PKIs from a system-wide installed
example). Fix this and install $EASYRSA_PKI/vars from a vars.example
(or just vars, as sometimes distributed) file elsewhere.