easyrsa: init-pki: fix bogus unset when printing the new PKI status
As written, select_vars() will never do anything after an init-pki
because $EASYRSA_NO_VARS will be set on the first run when entering
easyrsa. Unset that.

Additionally, do not unset $EASYRSA_VARS_FILE as it can be provided
externally (as an env var or as `--vars=` switch). By unsetting it,
we destroy the information about the fact we were passed an external
vars file. In the opposite case (vars copied from template), nothing
seems to rely on unsetting of this variable and the output seems
correct in both cases.