wrappers: -batman-preview: add a variant of `batman` for usage in vifm
This is another adapter for usage of `bat` as a prettifying `$MANPAGER`,
suited for usage as a `fileviewer` in vifm (i.e. output is to a pipe,
with vifm doing the actual paging). To accommodate this role, we force
ANSI output, disable paging and instead configure man and bat to use
vifm-provided preview area width for typesetting and wrapping.