wrappers: cargo: only bail on `$CARGO_PROFILE_*`
Do not trip over our own `$CARGO_CONFIG` (which is not even exported).
Regardless, given that the purpose of this wrapper is to provide
profile-specific configuration, only skip custom logic if _profiles_ are
being overridden by the user (and carry on if the user merely set some
variables affecting e.g. Cargo network behavior).

Fixes: fb2f734 ("wrappers: cargo: skip custom logic if `$RUSTFLAGS` or `$CARGO_*` is set")