util: disk-fstrim: slightly unbreak partition/disk determination
Apparently, PTTYPE is also set for individual partitions on a disk. Try to avoid
relying on that. `lsblk` provides a `TYPE` column that can be `disk` or `part`,
but unfortunately, it mixes up the notion of physical and logical types and this
column can also contain values such as `loop` or `crypt` (which can mean either
thing).

Try to invent a heuristic that somehow works.