util: sshfs-to: use SSH connection multiplexing
OK, so at first glance, this might look like an exercise in futility
because the entire idea of sshfs `-omax_conns` (which we also use) is
to use multiple connections for performance, and now we funnel them all
into the same fundamental TCP stream.

However, in practice this does not matter much, because multiple SFTP
connections are primarily used for I/O concurrency, not I/O parallelism,
and doing this does not really degrade performance. On the other hand,
this gets rid of the unpleasant effect where a running sshfs instance
may cause new SSH handshakes at arbitrary point in time (thus triggering
a hardware token interaction prompt, for instance).